handler fixes

+++ src/handler.ts
@@ -1,16 +1,17 @@
+import * as bytes from 'bytes';
 import * as http from 'http';
-import * as getRawBody from 'raw-body';
 import * as url from 'url';
 import {
   BaseHandler,
   BaseStorage,
+  DiskStorageConfig,
   ERRORS,
-  Request,
   File,
-  UploadXError,
+  Request,
   UploadxConfig,
-  DiskStorageConfig
+  UploadXError
 } from './core';
+import { getBody } from './utils';
 
 /**
  * X-headers  protocol implementation
@@ -20,6 +21,7 @@ export class Handler extends BaseHandler {
    * Where store files
    */
   storage: BaseStorage;
+  static idKey = 'upload_id';
 
   constructor(public options: UploadxConfig & DiskStorageConfig) {
     super();
@@ -28,65 +30,50 @@ export class Handler extends BaseHandler {
 
   /**
    * Build File from `create` request
-   * @internal
    */
-  private async buildFileFromRequest(req: Request) {
-    if (!req.body) {
-      try {
-        const raw = await getRawBody(req, {
-          length: req.headers['content-length'],
-          limit: '1mb',
-          encoding: true
-        });
-        req.body = JSON.parse(raw);
-      } catch (error) {
-        throw new UploadXError(ERRORS.INVALID_REQUEST, error);
-      }
-    }
+  protected async buildFileFromRequest(req: Request): Promise<File> {
     const file = {} as File;
-
-    file.metadata = req.body;
-    const user = req['user'];
-    file.userId = user && (user.id || user._id);
-    file.filename = req.body.name || req.body.title;
-    file.size = Number.parseInt(req.headers['x-upload-content-length'] || req.body.size);
-    file.mimeType = req.headers['x-upload-content-type'] || req.body.mimeType;
-
+    try {
+      file.metadata = await getBody(req);
+      'user' in req && (file.userId = req.user.id || req.user._id);
+      file.filename = file.metadata.name || file.metadata.title;
+      file.size = Number.parseInt(req.headers['x-upload-content-length'] || file.metadata.size);
+      file.mimeType = req.headers['x-upload-content-type'] || file.metadata.mimeType;
+    } catch (error) {
+      throw new UploadXError(ERRORS.BAD_REQUEST, error);
+    }
     if (!new RegExp((this.options.allowMIME || [`\/`]).join('|')).test(file.mimeType))
       throw new UploadXError(ERRORS.FILE_TYPE_NOT_ALLOWED);
     if (isNaN(file.size)) throw new UploadXError(ERRORS.INVALID_FILE_SIZE);
-    if (file.size > this.options.maxUploadSize!) throw new UploadXError(ERRORS.FILE_TOO_LARGE);
-    if (!file.filename) throw new UploadXError(ERRORS.INVALID_FILE_NAME);
-
+    if (file.size > this.options.maxUploadSize!)
+      throw new UploadXError(ERRORS.FILE_TOO_LARGE, `Max file size: ${this.options.maxUploadSize}`);
     return file;
   }
 
   /**
    * Build file url from request
-   * @internal
    */
-  private buildFileUrl(req: http.IncomingMessage, id: string) {
+  protected buildFileUrl(req: http.IncomingMessage, id: string): string {
     const urlObject = url.parse(req.url!, true);
     const query = urlObject.query;
-    const baseUrl = (req['baseUrl'] as string) || urlObject.pathname;
+    const baseUrl = (req['baseUrl'] as string) || urlObject.pathname || '';
     const search = Object.keys(query).reduce(
       (acc, key) => acc + `&${key}=${query[key] || ''}`,
-      `?upload_id=${id}`
+      `?${Handler.idKey}=${id}`
     );
     const location: string =
       !this.options.useRelativeURL && req.headers.host
-        ? `//${req.headers.host}${baseUrl || ''}${search}`
-        : `${baseUrl || ''}${search}`;
+        ? `//${req.headers.host}${baseUrl}${search}`
+        : `${baseUrl}${search}`;
     return location;
   }
 
   /**
    * Get `upload_id` from request
-   * @internal
    */
-  private getFileId(req: http.IncomingMessage) {
+  protected getFileId(req: http.IncomingMessage): string | undefined {
     const query = url.parse(req.url!, true).query;
-    return query && (query.upload_id as string);
+    return query && (query[Handler.idKey] as string);
   }
 
   /**
@@ -106,21 +93,22 @@ export class Handler extends BaseHandler {
   /**
    * Write chunk to file or/and return chunk offset
    */
-  async write(req: http.IncomingMessage, res: http.ServerResponse) {
+  async write(req: http.IncomingMessage, res: http.ServerResponse): Promise<File> {
     if (this.options.maxChunkSize && +req.headers['content-length']! > this.options.maxChunkSize) {
-      throw new UploadXError(ERRORS.CHUNK_TOO_BIG);
+      throw new UploadXError(
+        ERRORS.CHUNK_TOO_BIG,
+        `Chunk size limit: ${bytes(this.options.maxChunkSize as number)}`
+      );
     }
     const id = this.getFileId(req);
+    if (!id) throw new UploadXError(ERRORS.BAD_REQUEST, 'File id cannot be retrieved');
     const rangeHeader = req.headers['content-range'];
     if (!rangeHeader) throw new UploadXError(ERRORS.INVALID_RANGE);
-    const [total, end, start] = rangeHeader!
+    const [total, end, start] = rangeHeader
       .split(/\D+/)
       .filter(v => v.length)
       .map(s => +s)
       .reverse();
-    req.on('error', () => {
-      throw new UploadXError(ERRORS.UNKNOWN_ERROR);
-    });
     const file = await this.storage.write(req as any, { total, end, start, id });
     if (file.bytesWritten === file.size) {
       req['file'] = file;
@@ -129,6 +117,7 @@ export class Handler extends BaseHandler {
       res.setHeader('Range', `bytes=0-${file.bytesWritten! - 1}`);
       this.send(res, 308);
     }
+    return file;
   }
 
   list(req: http.IncomingMessage, res: http.ServerResponse) {
@@ -140,6 +129,7 @@ export class Handler extends BaseHandler {
    */
   delete(req: http.IncomingMessage, res: http.ServerResponse): Promise<File> {
     const id = this.getFileId(req);
+    if (!id) throw new UploadXError(ERRORS.BAD_REQUEST, 'File id cannot be retrieved');
     return this.storage.delete(id);
   }
 
@@ -150,7 +140,7 @@ export class Handler extends BaseHandler {
     const statusCode = error.statusCode || 500;
     const errorBody = {
       error: {
-        code: error.code,
+        code: error.code || 'unknown_error',
         message: error.message || 'unknown error'
       }
     };
